🚧 电源管理IC（drv）： 1、完成电源模块初始化时自动识别IC并初始化； 2、Tx、Rx能够正常编译通过； 结束

+++ applications/HL_A6902/driver/inc/hl_drv_sy6971.h
@@ -454,7 +454,7 @@ typedef struct _HL_SY6971_REGALL_T
     HL_SY6971_REG12_T reg12;
 }HL_SY6971_REGALL_T;
 
-uint8_t hl_drv_sy6971_init(void);
-uint8_t hl_drv_sy6971_io_ctrl(uint8_t cmd, void* ptr, uint8_t len);
+int hl_drv_sy6971_init(void);
+int hl_drv_sy6971_io_ctrl(uint8_t cmd, void* ptr, uint8_t len);
 
 #endif
